perf: pack storage slots

+++ src/lib/Structs.sol
@@ -49,12 +49,12 @@ struct Action {
 }
 
 struct RelativeStrategyConfig {
-  uint256 approvalPeriod; // The length of time of the approval period.
-  uint256 queuingPeriod; // The length of time of the queuing period. The disapproval period is the queuing period when
+  uint64 approvalPeriod; // The length of time of the approval period.
+  uint64 queuingPeriod; // The length of time of the queuing period. The disapproval period is the queuing period when
     // enabled.
-  uint256 expirationPeriod; // The length of time an action can be executed before it expires.
-  uint256 minApprovalPct; // Minimum percentage of total approval quantity / total approval supply.
-  uint256 minDisapprovalPct; // Minimum percentage of total disapproval quantity / total disapproval supply.
+  uint64 expirationPeriod; // The length of time an action can be executed before it expires.
+  uint16 minApprovalPct; // Minimum percentage of total approval quantity / total approval supply.
+  uint16 minDisapprovalPct; // Minimum percentage of total disapproval quantity / total disapproval supply.
   bool isFixedLengthApprovalPeriod; // Determines if an action be queued before approvalEndTime.
   uint8 approvalRole; // Anyone with this role can cast approval of an action.
   uint8 disapprovalRole; // Anyone with this role can cast disapproval of an action.
@@ -63,12 +63,12 @@ struct RelativeStrategyConfig {
 }
 
 struct AbsoluteStrategyConfig {
-  uint256 approvalPeriod; // The length of time of the approval period.
-  uint256 queuingPeriod; // The length of time of the queuing period. The disapproval period is the queuing period when
+  uint64 approvalPeriod; // The length of time of the approval period.
+  uint64 queuingPeriod; // The length of time of the queuing period. The disapproval period is the queuing period when
     // enabled.
-  uint256 expirationPeriod; // The length of time an action can be executed before it expires.
-  uint256 minApprovals; // Minimum number of total approval quantity.
-  uint256 minDisapprovals; // Minimum number of total disapproval quantity.
+  uint64 expirationPeriod; // The length of time an action can be executed before it expires.
+  uint128 minApprovals; // Minimum number of total approval quantity.
+  uint128 minDisapprovals; // Minimum number of total disapproval quantity.
   bool isFixedLengthApprovalPeriod; // Determines if an action be queued before approvalEndTime.
   uint8 approvalRole; // Anyone with this role can cast approval of an action.
   uint8 disapprovalRole; // Anyone with this role can cast disapproval of an action.
